Plot: update c_lib data types and progress percentage calculation
1. Change long to int64
2. Change all "count" variable to int64
3. Fix problematic progress percentage calculation


Former-commit-id: 1cd738ca4bd86f04847d5adcd0309f678eeb0d8e [formerly 0aec872790e955f88c5f75dab2e16abaf969982e]
Former-commit-id: 0c7410ff90598530dae4a3b892b8dc688f2ccf64

+++ gravity_plot/integrator_ias15.py
@@ -88,7 +88,7 @@ class IAS15:
         x_err_comp_sum = np.zeros((objects_count, 3))
         v_err_comp_sum = np.zeros((objects_count, 3))
 
-        count = ctypes.c_int(0)
+        count = ctypes.c_int64(0)
         store_count = ctypes.c_int(0)
         progress_bar = Progress_bar()
         with progress_bar:
